// types/lca.ts
export enum CaseStatus {
  CERTIFIED = 'Certified',
  DENIED = 'Denied',
  WITHDRAWN = 'Withdrawn',
  CERTIFIED_WITHDRAWN = 'Certified - Withdrawn'
}
// ...
export interface LCAData {
  lcaCaseNumber: string;
  applicationData: ApplicationData;
  jobData: JobData;
  employerData: EmployerData;
  legalData?: LegalData;
}

export interface JobData {
  jobTitle: string;
  socTitle: string;
  baseAnnualSalary: number;
}

export interface ApplicationData {
  caseNumber: string;
  caseStatus: CaseStatus;
  receivedDate: Date;
  decisionDate: Date;
}

export interface EmployerData {
  name: string;
  city: string;
  state: string;
}

export interface LegalData {
  attorneyName: string;
  attorneyFirmName?: string;
}